Validate Rails nested attributes
Consider the following example:

  class User < ActiveRecord::Base
    has_many :phone_numbers
    accepts_nested_attributes_for :phone_numbers
  end

  class PhoneNumber < ActiveRecord::Base
    # This is a bit contrived, because probably it should be
    # many-to-many
    has_many :labels
    accepts_nested_attributes_for :labels
  end

  class Label < ActiveRecord::Base
    belongs_to :phone_number
  end

If you render this on 1 page to edit labels you end up with the
following attributes name

  user[phone_numbers_attributes][0][labels_attributes][0][label]

And validation is on field
  user[phone_numbers_attributes][][label_attributes][][label]

At the moment the most current version of client side validation only
supports 1 level of nesting of accepts_nested_attributes_for. This
commit makes it support any level of nesting.